Move WebsiteUniqueConstraintService out of Events namespace
